Full compatibility with FCEUX 2.2.3
Note with FCEUX it doesn't store session data for resuming later, as FCEUX doesn't support persistent savestates properly.

+++ DankBot.lua
@@ -1,7 +1,9 @@
 --DankBot v2.1 by ThunderAxe31
+--works with BizHawk (not 2.5 & 2.5.1) and FCEUX 2.2.3
 --optimized for use on Gambatte with equal_length_frames set to False
 --place this script into its own folder, because it's going generate multiple files
---todo: implement automatic sanity check of route.lua; make it compatible with FCEUX;
+--make sure a proper route.lua file is placed in the same folder
+--todo: implement automatic sanity check of route.lua; implement multiple state add
 
 function table.val_to_str ( v )
   if "string" == type( v ) then
@@ -39,12 +41,46 @@ function table.tostring( tbl )
   return "{" .. table.concat( result, "," ) .. "}"
 end
 
-local function log_update(text, no_write_console)
+local text_output = 0
+if console and console.log then
+	text_output = console.log
+elseif emu and emu.print then
+	text_output = emu.print
+else
+	text_output = print
+end
+
+local pause = 0
+if client and client.pause then
+	pause = client.pause
+elseif emu and emu.pause then
+	pause = emu.pause
+end
+
+local unpause = 0
+if client and client.unpause then
+	unpause = client.unpause
+elseif emu and emu.unpause then
+	unpause = emu.unpause
+end
+
+local registerexit = 0
+if event and event.onexit then
+	registerexit = event.onexit
+elseif emu and emu.registerexit then
+	registerexit = emu.registerexit
+end
+
+local rerecordcounting = movie.setrerecordcounting or movie.rerecordcounting
+local movie_is_loaded = movie.isloaded or movie.active
+get_emu_time = emu.framecount() --this must stay global!
+
+local function log_update(text)
 	file_log = io.open("log.txt", "a")
 	io.output(file_log)
 	io.write(text .. "\n")
 	io.close(file_log)
-	console.log(text)
+	text_output(text)
 end
 
 check_same = true
@@ -58,8 +94,12 @@ function check_freeze(time_limit, check_type, check_value, func, argument)
 		count = count +1
 		if count > time_limit then
 			freezes = freezes +1
-			log_update("  WARNING: game appears to be frozen. Saving to freeze-" .. freezes .. ".bk2")
-			movie.save("freeze-" .. freezes)
+			if movie.save then
+				log_update("  WARNING: game appears to be frozen. Saving to freeze-" .. freezes .. ".bk2")
+				movie.save("freeze-" .. freezes)
+			else
+				log_update("  WARNING: game appears to be frozen.")
+			end
 			return false
 		end
 		emu.frameadvance()
@@ -81,7 +121,12 @@ state = {} --initializes the array of the savestate data
 state[0] = {}
 state[0][1] = {}
 state[0][1][1] = {}
-state[0][1][1]["slot"] = "0-1-1.State"
+if tastudio then --simple way of confirming if we're using BizHawk
+	state[0][1][1]["slot"] = "0-1-1.State"
+elseif taseditor then --otherwise, we'll just assume we're using FCEUX
+	state[0][1][1]["slot"] = savestate.object()
+	savestate.load(state[0][1][1]["slot"])
+end
 
 local current_action = 1
 
@@ -90,27 +135,39 @@ local function close_session()
 		log_update("Botting session SUSPENDED on " .. os.date("%y/%m/%d %X"))
 	end
 	movie.setreadonly(true)
-	movie.setrerecordcounting(true)--maybe this is just unnecessary and annoying, and movie.setreadonly()+savestate.load() does better the job
-	movie.stop()
-	memorysavestate.clearstatesfrommemory()--just in case
-	client.frameskip(0)
-	client.pause()
+	rerecordcounting(true)--maybe this is just unnecessary and annoying, and movie.setreadonly()+savestate.load() does better the job
+	if movie_is_loaded() then
+		movie.stop()
+	end
+	if memorysavestate and memorysavestate.clearstatesfrommemory then
+		memorysavestate.clearstatesfrommemory()--just in case
+	end
+	if client and client.frameskip then
+		client.frameskip(0)
+	end
+	pause()
 end
-event.onexit(close_session)
+if (tastudio == nil) and (taseditor == nil) then
+	text_output("ERROR: This script currently only works in BizHawk or FCEUX.")
+	current_action = #action+1
+end
+registerexit(close_session)
 
-local time_unit = ", cycle "
-client.unpause()
-local temp_state = memorysavestate.savecorestate()
-if emu.framecount() == 0 then emu.frameadvance() end
-get_emu_time = emu.totalexecutedcycles
-if emu.totalexecutedcycles() == 0 then
-	get_emu_time = emu.framecount
-	time_unit = ", frame "
+local time_unit = ", frame "
+get_emu_time = emu.framecount
+if emu.totalexecutedcycles then
+	unpause()
+	local temp_state = memorysavestate.savecorestate()
+	if emu.framecount() == 0 then emu.frameadvance() end
+	if emu.totalexecutedcycles() ~= 0 then
+		get_emu_time = emu.totalexecutedcycles
+		time_unit = ", cycle "
+	end
+	memorysavestate.loadcorestate(temp_state)
+	memorysavestate.removestate(temp_state)
+	console.clear()
+	pause()
 end
-memorysavestate.loadcorestate(temp_state)
-memorysavestate.removestate(temp_state)
-console.clear()
-client.pause()
 
 local function state_add(alt, wait, parent)
 	local cycle = get_emu_time()
@@ -145,8 +202,13 @@ local function state_add(alt, wait, parent)
 			for i=1, max_states do
 				if state[current_action][alt][i] == nil then
 					state[current_action][alt][i] = {}
-					state[current_action][alt][i]["slot"] = current_action .. "-" .. alt .. "-" .. i .. ".State"
-					savestate.save(state[current_action][alt][i]["slot"], true)
+					if tastudio then --simple way of confirming if we're using BizHawk
+						state[current_action][alt][i]["slot"] = current_action .. "-" .. alt .. "-" .. i .. ".State"
+						savestate.save(state[current_action][alt][i]["slot"], true)
+					else --otherwise, we'll just assume we're using FCEUX
+						state[current_action][alt][i]["slot"] = savestate.object()
+						savestate.save(state[current_action][alt][i]["slot"])
+					end
 					state[current_action][alt][i]["parent"] = parent --this is necessary for drawing the graphical output
 					state[current_action][alt][i]["wait"] = wait
 					state[current_action][alt][i]["cycle"] = cycle
@@ -163,7 +225,11 @@ local function state_add(alt, wait, parent)
 				end
 			end
 			if (state[current_action][alt][maximum_index]["cycle"] > cycle) then
-				savestate.save(state[current_action][alt][maximum_index]["slot"], true)
+				if tastudio then --simple way of confirming if we're using BizHawk
+					savestate.save(state[current_action][alt][maximum_index]["slot"], true)
+				else --otherwise, we'll just assume we're using FCEUX
+					savestate.save(state[current_action][alt][maximum_index]["slot"])
+				end
 				state[current_action][alt][maximum_index]["wait"] = wait
 				state[current_action][alt][maximum_index]["cycle"] = cycle
 				state[current_action][alt][maximum_index]["rng"] = rng
@@ -186,7 +252,11 @@ local function act(action_type, alts)
 				
 				log_update("  Looking from state " .. current_action-1 .. "-" .. alt_num .. "-" .. state_num)
 				
-				savestate.load(state[current_action-1][alt_num][state_num]["slot"], true)
+				if tastudio then --simple way of confirming if we're using BizHawk
+					savestate.load(state[current_action-1][alt_num][state_num]["slot"], true)
+				else --otherwise, we'll just assume we're using FCEUX
+					savestate.load(state[current_action-1][alt_num][state_num]["slot"])
+				end
 				
 				if action[current_action].func.prepare(alts[alt]) then
 					
@@ -207,10 +277,20 @@ local function act(action_type, alts)
 						wait_step = action[current_action][alt]["custom_wait_step"]
 					end
 					
-					local temp_state = memorysavestate.savecorestate()
+					local temp_state = 0
+					if memorysavestate and memorysavestate.clearstatesfrommemory then
+						temp_state = memorysavestate.savecorestate()
+					elseif taseditor then --otherwise, we'll just assume we're using FCEUX
+						temp_state = savestate.object()
+						savestate.save(temp_state)
+					end
 					
-					for wait=min_wait, max_wait, wait_step do						
-						memorysavestate.loadcorestate(temp_state)
+					for wait=min_wait, max_wait, wait_step do
+						if memorysavestate and memorysavestate.clearstatesfrommemory then
+							memorysavestate.loadcorestate(temp_state)
+						elseif taseditor then
+							savestate.load(temp_state)
+						end
 						
 						for i=1, wait do--we wait for a given amount of frames
 							emu.frameadvance()
@@ -236,8 +316,11 @@ local function act(action_type, alts)
 							state_add(alt, wait, {alt_num, state_num})
 						end
 					end
-					
-					memorysavestate.removestate(temp_state)
+					if memorysavestate and memorysavestate.clearstatesfrommemory then
+						memorysavestate.removestate(temp_state)
+					elseif taseditor then
+						temp_state = nil --dunno if it helps clearing memory
+					end
 				end
 			end
 		end
@@ -247,13 +330,38 @@ local function act(action_type, alts)
 	end
 end
 
-if not movie.isloaded() then --since there is no function to create a new movie, the user must open a movie before launching the script
+if not movie_is_loaded() then --since there is no function to create a new movie, the user must open a movie before launching the script
 	current_action = false
-	console.log("ERROR: No movie file loaded. IMPORTANT NOTE: The movie file that it's loaded before launching the script will be used as base for the botting session, so ALL data previously contained will be discarded. For this reason, it is recommended to load an empty movie or an otherwise disposable movie file, for example a copy.")
+	text_output("ERROR: No movie file loaded. IMPORTANT NOTE: The movie file that it's loaded before launching the script will be used as base for the botting session, so ALL data previously contained will be discarded. For this reason, it is recommended to load an empty movie or an otherwise disposable movie file, for example a copy.")
 	current_action = #action+1
 else
 	local file_resume = io.open("resume.lua", "r")
-	if file_resume then --this block checks for resume.lua and resumes a suspended session
+	if taseditor or (file_resume == nil) then --if using FCEUX or resume.lua doesn't exist, wipe log.txt and start a new session from scratch
+		local file_log = io.open("log.txt", "w")
+		io.output(file_log)
+		io.write("")
+		io.close(file_log)
+		
+		if tastudio then --simple way of confirming if we're using BizHawk
+			savestate.save(state[0][1][1]["slot"], true)
+		else --otherwise, we'll just assume we're using FCEUX
+			savestate.save(state[0][1][1]["slot"])
+		end
+		
+		local cycle = get_emu_time()
+		state[0][1][1]["wait"] = 0
+		state[0][1][1]["cycle"] = cycle
+		state[0][1][1]["rng"] = get_rng()
+		if console then
+			console.clear()
+		end
+		log_update("DankBot v2.1 by ThunderAxe31, session STARTED")
+		local rng_display = ""
+		if get_rng then
+			rng_display = ", RNG " .. string.format("%X", get_rng())
+		end
+		log_update(" Initializing state 0-1-1" .. time_unit .. cycle .. rng_display)
+	else --if we're using BizHawk and resume.lua exists, resume the suspended session
 		io.input(file_resume)
 		local file_resume_data = io.read()
 		io.close(file_resume)
@@ -268,51 +376,36 @@ else
 					resume_data["restart_action"] = nil
 					state = resume_data
 					if current_action == #action+1 then
-						console.log("Previous session is already complete. Add more actions or delete resume.lua")
+						text_output("Previous session is already complete. Add more actions or delete resume.lua")
 					elseif current_action > #action+1 then
-						console.log("ERROR: Session from resume.lua contains more actions than route.lua")
+						text_output("ERROR: Session from resume.lua contains more actions than route.lua")
 					else
 						log_update("Botting session RESUMED   on " .. os.date("%y/%m/%d %X"))
 					end
 				else
-					console.log("ERROR: Invalid resume data found in resume.lua")
+					text_output("ERROR: Invalid resume data found in resume.lua")
 					current_action = #action+1
 				end
 			else
-				console.log("ERROR: Invalid resume data found in resume.lua")
+				text_output("ERROR: Invalid resume data found in resume.lua")
 				current_action = #action+1
 			end
 		else
-			console.log("ERROR: Unable to read resume.lua")
+			text_output("ERROR: Unable to read resume.lua")
 			current_action = #action+1
 		end
-	else --if resume.lua doesn't exist, wipe log.txt and start a new session from scratch
-		local file_log = io.open("log.txt", "w")
-		io.output(file_log)
-		io.write("")
-		io.close(file_log)
-		savestate.save(state[0][1][1]["slot"], true)
-		local cycle = get_emu_time()
-		state[0][1][1]["wait"] = 0
-		state[0][1][1]["cycle"] = cycle
-		state[0][1][1]["rng"] = get_rng()
-		console.clear()
-		log_update("DankBot v2.1 by ThunderAxe31, session STARTED")
-		local rng_display = ""
-		if get_rng then
-			rng_display = ", RNG " .. string.format("%X", get_rng())
-		end
-		log_update(" Initializing state 0-1-1" .. time_unit .. cycle .. rng_display)
 	end
 	if get_rng == nil then
 		log_update("WARNING: get_rng() function not declared in route.lua")
 	end
 end
 
-movie.setrerecordcounting(false)
+rerecordcounting(false)
 movie.setreadonly(false)
-client.frameskip(9)
-client.unpause()
+if client and client.frameskip then
+	client.frameskip(9)
+end
+unpause()
 
 while current_action <= #action do--this is the main code block that controls the botting flow
 	state[current_action] = {}
@@ -339,7 +432,7 @@ while current_action <= #action do--this is the main code block that controls th
 		log_update("Botting session FAILED on " .. os.date("%y/%m/%d %X"))
 		current_action = #action+1 --this will prevent to display "botting session suspended" message
 	else
-		if not movie.isloaded() then --we don't want to store savestates that don't contain movie inputs, so let's not save this action data
+		if not movie_is_loaded() then --we don't want to store savestates that don't contain movie inputs, so let's not save this action data
 			log_update(" ERROR: It appears you closed the movie file.")
 			log_update("Botting session INTERRUPTED on " .. os.date("%y/%m/%d %X"))
 			current_action = #action+1
@@ -364,22 +457,36 @@ while current_action <= #action do--this is the main code block that controls th
 					end
 				end
 				log_update(" Closing with state " .. current_action .. "-" .. minimum_alt .. "-" .. minumum_index .. time_unit .. state[current_action][minimum_alt][minumum_index]["cycle"])
-				savestate.load(state[current_action][minimum_alt][minumum_index]["slot"], true)
+				
+				if tastudio then --simple way of confirming if we're using BizHawk
+					savestate.load(state[current_action][minimum_alt][minumum_index]["slot"], true)
+				else --otherwise, we'll just assume we're using FCEUX
+					savestate.load(state[current_action][minimum_alt][minumum_index]["slot"])
+				end
+				
 				log_update("Botting session COMPLETED on " .. os.date("%y/%m/%d %X"))
 			end
 			
-			for z=1, #state[current_action-1] do --this deletes the old savestate files from disk
-				for i=1, #state[current_action-1][z] do
-					os.remove(state[current_action-1][z][i]["slot"])
-					state[current_action-1][z][i]["slot"] = nil
+			if tastudio then
+				for z=1, #state[current_action-1] do --this deletes the old savestate files from disk
+					for i=1, #state[current_action-1][z] do
+						os.remove(state[current_action-1][z][i]["slot"])
+						state[current_action-1][z][i]["slot"] = nil
+					end
+				end
+				
+				state["restart_action"] = current_action +1
+				file_resume = io.open("resume.lua", "w")
+				io.output(file_resume)
+				io.write("return " .. table.tostring(state)) --save state data for resuming
+				io.close(file_resume)
+			elseif taseditor then
+				for z=1, #state[current_action-1] do
+					for i=1, #state[current_action-1][z] do
+						state[current_action-1][z][i]["slot"] = nil
+					end
 				end
 			end
-			
-			state["restart_action"] = current_action +1
-			file_resume = io.open("resume.lua", "w")
-			io.output(file_resume)
-			io.write("return " .. table.tostring(state)) --save state data for resuming
-			io.close(file_resume)
 			
 			if(current_action>=2) then
 				if export_treeview then
@@ -406,4 +513,8 @@ while current_action <= #action do--this is the main code block that controls th
 			current_action = current_action +1
 		end
 	end
+end
+
+if taseditor then
+	close_session() --because FCEUX doesn't consider a script as closed if it merely reaches the end
 end
